+++ GW2EIEvtcParser/EncounterLogic/Raids/W5/StatueOfDarkness.cs
@@ -86,7 +86,7 @@ namespace GW2EIEvtcParser.EncounterLogic
             CombatItem logStartNPCUpdate = combatData.FirstOrDefault(x => x.IsStateChange == ArcDPSEnums.StateChange.LogStartNPCUpdate);
             if (logStartNPCUpdate != null)
             {
-                IReadOnlyList<AgentItem> lightThieves = agentData.GetNPCsByID((int)ArcDPSEnums.TrashID.LightThieves);
+                IReadOnlyList<AgentItem> lightThieves = agentData.GetNPCsByID(ArcDPSEnums.TrashID.LightThieves);
                 if (lightThieves.Any())
                 {
                     startToUse = lightThieves.Min(x => x.FirstAware);
